requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Werkzeug==3.1.3
libsass==0.23.0
Flask-Scss==0.5
gunicorn==23.0.0
pre-commit==4.3.0

# If you use docker
PyMySQL==1.1.2
cryptography==46.0.2